+++ src/common/common_properties.graphql
