+++ src/common/controller_common_functions.js
